Move rounding logic in gcodegenerator
